requirements.txt
discord.py
python-dotenv
oandapyV20
gspread
google-oauth2-service-account
pandas
